Update Exection of The Fastbin
This code is an exploit script written in Python using the pwntools library. It targets a binary named fastbin_dup to exploit a heap-based vulnerability known as "fastbin double-free." The goal of the exploit is to manipulate the heap and eventually gain arbitrary code execution by overwriting the __malloc_hook function pointer in the GNU C Library (libc) with the address of a one-gadget, which will trigger a shell when executed.